create proxmox-rrd-migration-tool
This tool is intended to migrate the Proxmox VE (PVE) RRD data files to
the new schema.

Up until PVE8 the schema has been the same for a long time. With PVE9 we
introduced new columns to guests (vm) and nodes. We also switched all
types (vm, node, storate) to the same aggregation schemas as we do it in
PBS.
The result of both are a much finer resolution for long time spans, but
also larger RRD files.

* node: 79K -> 1.4M
* vm: 66K -> 1.3m
* storage: 14K -> 156K

The old directories for VMs used to be in `/var/lib/rrdcached/db/` with
the following sub directories:

* nodes: `pve2-node`
* guests (VM/CT): `pve2-vm`
* storage: `pve2-storage`

With this change we also introduce a new key schema, that makes it
easier in the future to introduce new ones. Instead of the
`pve{version}-{type}` we are switching to `pve-{type}-{version}`.

This enables us to add new columns with a new version, without breaking
nodes that are not yet updated. We are NOT allowed to remove or re-use
existing columns. That would be a breaking change.
We are currently at version 9.0. But in the future, if needed, this tool
can be adapted to do other migrations too.
For example, {old, 9.0} -> 9.2, should that be necessary.

The actual migration is handled by `librrd` to which we pass the path to
the old and new files, and the new RRD definitions. The `rrd_create_r2`
call then does the hard work of migrating and converting exisiting data
into the new file and aggregation schema.

This can take some time. Quick tests on a Ryzen 7900X with the following
files:
* 1 node RRD file
* 10k vm RRD files
* 1 storage RRD file

showed the folling results:

* 1 thread:  179.61s user 14.82s system 100% cpu 3:14.17 total
* 4 threads: 187.57s user 16.98s system 399% cpu 51.198 total

This is why we do not migrate inline, but have it as a separate step
during package upgrades.

Behavior: By default nothing will be changed and a dry or test run will
happen.
Only if the `--migrate` parameter is added will the actual migration be
done.

For each found RRD source file, the tool checks if a matching target
file already exists. By default, those will be skipped to not overwrite
target files that might already store newer data.
With the `--force` parameter this can be changed.

That means, one can run the tool multiple times (without --force) and it
will pick up where it might have left off. For example it the migration
was interrupted for some reason.

Once a source file has been processed it will be renamed with the `.old`
appendix. It will be excluded from future runs as we check for files
without an extension.

The tool has some simple heuristic to determine how many threads should
be used. Be default the range is between 1 to 4 threads. But the
`--threads` parameter allows a manual override.

Signed-off-by: Aaron Lauterer <a.lauterer@proxmox.com>