output errors from migration and mention the subject in the message
Avoid the not-so-telling "would migrate but in dry run mode" times
the count of resources but rather print the "error" message, which
includes not migrating due to dry-run mode or target already existing.

Use a separate message for dry-rune mode and existing target for
clarity and include the ID of the subject we'd currently migrate the
metrics for, that way it's much more telling compared the same string
X times.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>