add guarded systemd service for handling the migration on-boot
The startup of the unit file is guarded by a condition that the path
/var/lib/pve-manager/on-boot-rrd-migration-trigger exists. That way we
can touch this path in the manager on upgrade and trigger a migration
on the next boot that will definitively happen before pmxcfs,
rrdcached  and pvestatd start up and thus can write (potentially old)
data.

As it's guarded and thus opt-in we can still easily move to another
mechanism for doing the migrate, e.g. for setups with a huge amount of
guests (and storage * node for that matter) can be nudged towards
triggering the migration manually upfront.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>