fix failure during migration when there are no guest metrics yet
If a node never has had any guests on it, the pve2-vm directory does not
exist yet, which led to a failure when attempting to read the contents
of this directory. This is fixed by explicitly checking for a 'not
found' error (we still want to fail for other kinds of errors).

Signed-off-by: Lukas Wagner <l.wagner@proxmox.com>
Link: https://lore.proxmox.com/20250731082847.1113092-1-l.wagner@proxmox.com