collect rrd files: only skip '.old' extension to better handle storage IDs containing a dot
Storage IDs may contain a dot, which would make the tool ignore data
for them. This does not fix the issue completely, the final corner
case with storages being called called 'something.old' remains.

A warning for storages called '.old' is added to the pve8to9 checker
script. If enough people really run into this issue with such
storages, the migration script will be improved further after the
release of PVE 9. For now, use a straight-forward stop-gap measure to
avoid the fallout from not handling any storage ID with a dot.

Signed-off-by: Fiona Ebner <f.ebner@proxmox.com>
Link: https://lore.proxmox.com/20250804125601.91944-2-f.ebner@proxmox.com